Split sidebar into agent and builder view
